+++ src/main/java/com/joinai_support/repository/SupportTicketRepository.java
@@ -3,5 +3,5 @@ package com.joinai_support.repository;
 import com.joinai_support.domain.SupportTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupportTicketRepository extends JpaRepository<SupportTicket, Integer> {
+public interface SupportTicketRepository extends JpaRepository<SupportTicket, Long> {
 }
